Lab10 - Working with Docker Containers/shopease/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.21
python-dotenv==1.0.0
psycopg2-binary==2.9.7
